refactor: 컨트롤러 연결
