+++ src/main/java/kitchenpos/products/tobe/application/DefaultProductService.kt
@@ -1,22 +1,24 @@
 package kitchenpos.products.tobe.application
 
 import kitchenpos.menus.tode.application.PriceChangeDrivenMenuUpdater
+import kitchenpos.products.application.ProductService
 import kitchenpos.products.domain.Product
 import kitchenpos.products.domain.ProductRepository
 import kitchenpos.products.tobe.domain.CreateProductService
+import org.springframework.core.annotation.Order
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.util.*
 
 @Transactional
-@Service
+@Service("ProductService")
 class DefaultProductService(
     private val productRepository: ProductRepository,
     private val createProductService: CreateProductService,
     private val changeProductPriceService: ChangeProductPriceService,
     private val priceChangeDrivenMenuUpdater: PriceChangeDrivenMenuUpdater,
-) {
-    fun create(request: Product): Product {
+): ProductService {
+    override fun create(request: Product): Product {
         val product = createProductService.createProduct(
             price = request.price,
             name = request.name,
@@ -25,7 +27,7 @@ class DefaultProductService(
         return productRepository.save(product)
     }
 
-    fun changePrice(
+    override fun changePrice(
         productId: UUID,
         request: Product,
     ): Product {
@@ -40,5 +42,5 @@ class DefaultProductService(
     }
 
     @Transactional(readOnly = true)
-    fun findAll(): List<Product> = productRepository.findAll()
+    override fun findAll(): List<Product> = productRepository.findAll()
 }
